Work around special cases for ** token
PtrTypeStart parsing is made more complicated than it seems
because of the fact that ** is a single token. The accomodations
made because of this are somewhat brittle and unintuitive. An
alternative solution would be to let each asterisk stay its own
token, and do the extra work when expecting a **. However, this
would require a change to the grammar and would make different
spacing legal, e.g. `"abcd" * * 4`.