Fixes #2379
There were a few UTF-8 edge cases not covered, which this patch
covers.

The octets C0 and C1 are universally invalid according to UTF-8. Why?

1. As solo bytes, they are outside of the range of ASCII (00-7F).
2. As continuation bytes, they do not start with the bit pattern 0b10.
3. As leading bytes of a multibyte sequence, they would result in a
   2-byte "overlong" representation of ASCII which could have been
   encoded as plain ASCII. UTF-8 disallows this inferior encoding.

This patch fixes the third case above; the others were already covered.

Next, the octets F5 through FF are never seen in UTF-8. Why?

1. Same as above!
2. Same as above!
3. As leading bytes of a multibyte sequence, they would result in a
   4-byte sequence which encodes a codepoint beyond U+10FFFF. That
   is the highest codepoint supported by UTF-16, and UTF-8 explicitly
   limits its encoding range to that of UTF-16.

Again, this patch fixes the third case.

References used for this change:

- https://tools.ietf.org/html/rfc3629
- https://tools.ietf.org/html/rfc8259

This patch explicitly ignores the case of (technically) disallowed
"surrogate pairs" bytes in the range D800 to DFFF because the Zig
community already has ongoing discussions about how to handle that.